refactor(scratchpad): split solver modules by concern; drop dead code
Reorganize the permutation-based layout solver for reviewability, with no
behavior change on the production path:

- Extract the `Profile` contact-profile data structure into
  `contact_profile.py` and the cooling schedules into `cooling_schedules.py`,
  so `permutation_layout.py` and `imanishi_xu.py` each cover a single concern.
- Remove the unused `CoolingScheduleFromPaper` schedule and its "from_paper"
  option (unreached by any test, benchmark, or doc).
- Drop the production-only "rebuild" rotate-profile mode; its from-scratch
  correctness oracle now lives entirely in the differential test suite
  (`FastRotateTests` already compares against a freshly-built solver).
- Note `contact_at` as inspection/test-only (the hot path inlines it).
- Rewrite the design doc as current-state documentation, dropping the
  chronological build narrative and the PR-era open-questions section.

Update all import sites (tests, benchmarks, examples) to the new modules.
