+++ docs/source/user_guide/examples/scratchpad/inplace_annealing.py
@@ -36,8 +36,10 @@ from collections import Counter
 import torch  # noqa: F401
 
 from torch_spyre._inductor.scratchpad.plan_solver import LifetimeBoundBuffer
-from torch_spyre._inductor.scratchpad.imanishi_xu import (
+from torch_spyre._inductor.scratchpad.cooling_schedules import (
     ExponentialCoolingSchedule,
+)
+from torch_spyre._inductor.scratchpad.imanishi_xu import (
     ImanishiXuSolverWithBuffers,
 )
 from torch_spyre._inductor.scratchpad.utils import plot_buffers, quality_plot
